[Design Pattern] Add Observer Pattern
This implements the proposed AnnouncementManager <> Student observer
pattern proposed in our initial implementation spec for our Course
Management System.

When a teacher creates an announcement for a course, the
announcement_manager will have its notify method called which sends an
update message to each of its subscribes (students). Essentially,
students will be able to be notified on a per course basis on any
important updates by the professor.