[Functionality] Implement basic login with bcrypt
Using the bcrypt gem, this adds basic sessions, helper methods for
authorization, logout, and logged_in, and a basic ui login and
registration page for our Course Management System. Once users are
logged in, they can register to create a student or teacher account
which will grant them access to the rest of the functionality of our
Course Management System.