[Functionality] Add UserRegistration model
This adds the UserRegistration model which represents a registration
request placed by a user in our course management system. A user can
make multiple registration requests at a time for either a Student,
Teacher, or Admin account type which an Admin user can then approve. The
status of the registration can either be pending, approved, or denied.