[Functionality] UI updates, validations, and tweaks
This updates all forms to be more bootstrap friendly, changes a few
relationships and validations on models, passes in a new type field
on the user form, re-styles the flash error, redirects for the flash
error handling for User object create, adds a rescue NoMethodError on
the full_name User method, and adds a basic seed task for testing.

Multiple minor changes were encompassed in this commit as a preparatory
step for our registration approval flow.