[Functionality] Add admin course registration actions
This adds the approve and deny course registration actions for admin
users on the root page. Once admins are logged in, they will now be able
to view all pending course registration requests and select a button to
either approve or deny their request. This will be updated in real-time
on the student dashboard.