[Refactoring] Move helpers out of application_controller
Having helper_method in application_controller is a bit of a code smell
following Rails conventions that we want to avoid in the long run. This
moves all of the helper methods into the ApplicationHelper and includes
it in the controller instead.

Also replaces some of the unnecessary helpers with the direct reference
and deletes the empty helpers.