[Refactoring] Layout restyling and work tweak
This restyles the css property of link to stay blue after click so that
on page reload, the link stays blue instead of greyed out. Also restyles
the h1 header in the home page to be of nil class so that the text font
will be folder. Finally, this touches up the card header wordings to a
different possessive determiner so that, as a user, the system makes
more sense to me.