[Functionality] Add has_many scope and model helpers
This adds the where -> status: "approved" for both the course and user
model on the course_registrations relationship so that denied and
pending requests are filtered out (only one approved per User<>Course).

Also adds a mocked final_grade helper for the CourseRegistration model
and an active_course_registration helper on the Student model so that
finding the active CourseRegistration is easier.