[Design Pattern]: Prepare Abstract Factory Pattern
This makes the relevant schema changes, updates the abstract_factory
diagram, and unsets the abstract_class=true in the models for
Deliverable so that we're ready to implement a working Abstract Factory
Pattern. Using a combination of multiple and single table inheritance,
we set Activity to be an abstract class (MTI) then have our subclasses
Deliverable and Resource be concrete and store the subtypes as a column.