[Functionality] Add deliverable/resource controllers (STI)
This adds the deliverable and resource controllers along with their
respective CRUD actions to allow for the creation, updating, reading and
deleting of deliverables/resources by authorized users. By Rails' STI
conventions the subclass controllers are delegated to the superclass on
the basis that there exists a type column on the parent class table
model.

Additionally includes the relevant views corresponding to the new, edit,
and show actions.