[Functionality] Add param sanitization on STI controllers
One caveat of using Single Table Inheritance in Rails is that the
specific subtype of object will be lost on edit/update (the default
parent class will always be used) meaning that the param is missing
exception will raise every time. This adds a before_action callback that
catches this behaviour by verifying if a key with value exists already
in the params and replaces the default parent class hash values with
that value. Essentially fixes the update functionality for all users,
resources, and deliverables.