[Functionality] Hotfix user registration
The previous helper methods in the course.rb model for checking
registered? and pending_course_registration? were not working as
intended since they were missing the student_id param that was mandatory
per CourseRegistration filtration. Without this, the `Register` button
would not show if there were any outstanding pending registration
requests on any user. This new logic applies the hotfix by passing in
the id of the current user.